A very basic implementation of loco function configuration and activation (#30)
Added new page to train editor to configure locomotive functions.
Function config is saved to the the loco's settings in the database.
WebSocket API handler for function activation requests.
Added new page for more detailed locomotive control.

The UI for this is really rough and needs to be improved. The main goal with this revision is to test the core functionality then then figure out the best way to present it to the user.