Added support for variables in CLI scripts (#57)
Added support for variables in command scripts. The variables are stored in the command context and can be accessed using the following:
* 'set' command to set a variable
* 'unset' to clear a variable
* '$VARNAME' to use a variable in a command, e.g. 'loco_speed $locoId 64'

Added support for line commends using '#'.
Fixed some parsing issues related to quoted empty strings.
Split command handling out into its own module so that it can be used outside of the CLI in the future.
Bumped version number to 0.3.2.